Implementation of alternate "inconsistent" read consistency.
MVCCGet and MVCCScan now take a "consistent" boolean which, if false,
ignores intents. Moved the guts of MVCCScan into MVCCIterate and MVCCScan
just calls into it. This change allows the removal of the MVCCIterateCommitted
method.

Provides a unittest for "deadloop" issue of encountering an extant intent
at a meta indexing record.

Fixes #343
Fixes #435