Handle interaction of read consistency with multi-range ops.
If read consistency isn't required, a timestamp is assigned using
the local clock. This prevents cases where a different timestamp is
used for each part of a multi-range op, each assigned to the
respective node's clock at the time the RPC is received.

Also, if read consistency isn't required, don't insist that ops
spanning ranges be auto-wrapped in a transaction.

Fixed some errors in dist_sender.go with error handling. In one case
the error wasn't being cleared on the retry loop. In another, the
response error wasn't being examined after calling the RPC.